+++ crates/burn-jit/src/kernel/matmul/tiling2d_shader/mod.rs
@@ -1,0 +1,11 @@
+mod base;
+mod computation;
+mod load_shared_memory;
+mod shader_information;
+mod write_output;
+
+pub(crate) use base::*;
+pub(crate) use computation::*;
+pub(crate) use load_shared_memory::*;
+pub(crate) use shader_information::*;
+pub(crate) use write_output::*;
